Delete project link

+++ src/api/controllers/projectController.py
@@ -110,6 +110,7 @@ class ProjectController:
             if "link" in kwargs and "project_id" in kwargs and "name" in kwargs["link"] and "url" in kwargs["link"] and len(kwargs) == 2:
                 if kwargs["link"]["name"] is None or kwargs["link"]["url"] is None or kwargs["project_id"] is None:
                     return None, "Arguments can't be empty", 400
+
                 link = ProjectLink(project_id=kwargs["project_id"], **kwargs["link"])
                 self.session.add(link)
                 self.session.commit()
@@ -122,12 +123,10 @@ class ProjectController:
             return None, "Failed to create project link.", 400
 
     def update_link(self, user_id, **kwargs):
-        if "project_id" not in kwargs:
-            return None, "Failed to update project link. Request body must specify link's 'project_id'.", 400
-
-        if "link" in kwargs and "project_id" in kwargs and "id" in kwargs["link"] and len(kwargs) == 2:
-            if kwargs["link"]["id"] is None or kwargs["project_id"] is None:
+        if "project_id" in kwargs and "link" in kwargs and "id" in kwargs["link"] and len(kwargs) == 2:
+            if kwargs["project_id"] is None or kwargs["link"]["id"] is None:
                 return None, "Arguments can't be empty", 400
+
             link = ProjectLink.query.filter_by(project_id=kwargs["project_id"], id=kwargs["link"]["id"]).first()
 
             if link == None:
@@ -151,16 +150,22 @@ class ProjectController:
 
         return all_links
 
-    def delete_link(self, project_id, link_id):
-        link = ProjectLink.query.filter_by(project_id=project_id, id=link_id).first()
+    def delete_link(self, user_id, **kwargs):
+        if "project_id" in kwargs and "link" in kwargs and "id" in kwargs["link"] and len(kwargs) == 2 and len(kwargs["link"]) == 1:
+            if kwargs["project_id"] is None or kwargs["link"]["id"] is None:
+                return None, "Arguments can't be empty", 400
 
-        if link == None:
-            return None
+            link = ProjectLink.query.filter_by(project_id=kwargs["project_id"], id=kwargs["link"]["id"]).first()
 
-        db.session.delete(link)
-        db.session.commit()
+            if link == None:
+                return None, "Failed to delete project link. Project link not found.", 404
 
-        return link
+            db.session.delete(link)
+            db.session.commit()
+
+            return link, "OK", 200
+        else:
+            return None, "Forbidden attributes used in request. Only 'project_id' and 'link' object containing 'id' allowed.", 400
 
     # Project Feedback
     def create_feedback(self, project_id, **kwargs):
